+++ src/lib/vector_search_index_updater.py
@@ -27,17 +27,21 @@ class VectorSearchIndexUpdater:
         try:
             client = MongoClient(mongo_cluster_uri)
             db = client.get_database()
-            self.collection = db[collection_name]
         # pylint: disable=broad-except
         except Exception:
-            self.collection = client[db_name][collection_name]
+            db = client[db_name]
+
+        # Create the collection if it does not exist
+        if collection_name not in db.list_collection_names():
+            self.logger.info(f'Collection "{collection_name}" missing, it will be created now')
+            db.create_collection(collection_name)
+        self.collection = db[collection_name]
 
     def _create_vector_index(self, new_index_definition: SearchIndexModel) -> None:
-        self.logger.info(f'Vector Search index "{self.index_name}" missing, it will be created now')
+        self.logger.info(f'Vector Search index "{self.index_name}" missing in {self.collection.name}, it will be created now')
         self.logger.debug(f'Creating Vector Search index to the following definition: "{new_index_definition.document.get("definition")}"')
         self.collection.create_search_index(model=new_index_definition)
         self.logger.info(f'Created Vector Search index "{self.index_name}"')
-
 
     def _update_vector_index(self, current_index_definition: SearchIndexModel, new_index_definition: SearchIndexModel):
         self.logger.info(f'Check of MongoDB Vector Search index "{self.index_name}"')
@@ -51,8 +55,7 @@ class VectorSearchIndexUpdater:
 
         self.logger.debug(f'Updating Vector Search index to the following definition: "{new_fields.get("fields")}"')
         self.collection.update_search_index(self.index_name, new_fields)
-        self.logger.info(f'Updated Vector Search index "{self.index_name}"')
-
+        self.logger.info(f'Updated Vector Search index "{self.index_name}" in collection {self.collection.name}')
 
     def _get_current_vector_index_definition(self) -> SearchIndexModel | None:
         indexes = self.collection.list_search_indexes()
@@ -63,7 +66,6 @@ class VectorSearchIndexUpdater:
             return None
 
         return SearchIndexModel(definition=vector_index.get("latestDefinition"), name=vector_index.get("name"), type="vectorSearch")
-
 
     def _get_updated_vector_index_definition(self) -> SearchIndexModel:
         configured_similarity_fn = self.app_context.configurations.vectorStore.relevanceScoreFn or RelevanceScoreFn.cosine
@@ -84,7 +86,6 @@ class VectorSearchIndexUpdater:
             type=VECTOR_INDEX_TYPE,
         )
 
-
     def update_vector_search_index(self) -> None:
         try:
             current_vector_index_definition = self._get_current_vector_index_definition()
@@ -96,6 +97,6 @@ class VectorSearchIndexUpdater:
                 self._update_vector_index(current_vector_index_definition, new_vector_index_definition)
         # pylint: disable=broad-except
         except Exception as ex:
-            self.logger.warn(f'Unable to update Vector Search index "{self.index_name}".')
-            self.logger.warn(ex)
-            self.logger.warn("Service will continue to run, but you might experience unwanted behaviors.")
+            self.logger.warning(f'Unable to update Vector Search index "{self.index_name}".')
+            self.logger.warning(ex)
+            self.logger.warning("Service will continue to run, but you might experience unwanted behaviors.")