+++ tests/src/api/controllers/embedding_handler_test.py
@@ -39,7 +39,7 @@ def test_generate_embeddings_from_url_conflict(test_client):
 )
 def test_generate_embeddings_from_file(test_client, file_name, file_content, content_type):
     with patch("src.api.controllers.embeddings.embeddings_handler.EmbeddingGenerator.generate_from_text") as mock_generate_from_text, \
-        patch("src.application.embeddings.file_parser.FileParser.extract_documents_from_file") as mock_extract_documents_from_file:
+        patch("src.application.embeddings.file_parser.file_parser.FileParser.extract_documents_from_file") as mock_extract_documents_from_file:
         mock_extract_documents_from_file.return_value = ["Mock content"]
 
         file = (file_name, file_content, content_type)
@@ -60,7 +60,7 @@ def test_generate_embeddings_from_zip_file(test_client):
     buffer.seek(0)
 
     with patch("src.api.controllers.embeddings.embeddings_handler.EmbeddingGenerator.generate_from_text") as mock_generate_from_text, \
-        patch("src.application.embeddings.file_parser.FileParser.extract_documents_from_file") as mock_extract_documents_from_file:
+        patch("src.application.embeddings.file_parser.file_parser.FileParser.extract_documents_from_file") as mock_extract_documents_from_file:
 
         mock_extract_documents_from_file.return_value = ["This is a text file", "This is a markdown file"]
         files = {"file": ("zip_file.zip", buffer, "application/zip")}
@@ -74,18 +74,6 @@ def test_generate_embeddings_from_zip_file(test_client):
 
         mock_generate_from_text.assert_any_call("This is a text file")
         mock_generate_from_text.assert_any_call("This is a markdown file")
-
-
-def test_fail_generate_embeddings_from_unsupported_file(test_client):
-    with patch("src.api.controllers.embeddings.embeddings_handler.EmbeddingGenerator.generate_from_text") as mock_generate_from_text, \
-        patch("src.application.embeddings.file_parser.FileParser.extract_documents_from_file") as mock_extract_documents_from_file:
-        files = {"file": ("unsupported_file.wrong", b"Content of an unsupported file", "application/octet-stream")}
-        response = test_client.post("/embeddings/generateFromFile", files=files)
-        
-        assert response.status_code == 400
-        assert response.json() == {"detail": "Application does not support this file type (content type: application/octet-stream)."}
-        mock_extract_documents_from_file.assert_not_called()
-        mock_generate_from_text.assert_not_called()
 
 @pytest.mark.parametrize(
     "file_name, content_type",